[multibody] Refile icf_model and constraint tests
icf_model and all of the *_constraints_pool modules are in a link-time
dependency cycle. Perhaps naturally, this has led to piling almost all of the
relevant tests for the whole group into icf_model_test.cc.

In order to make room for quite a few more tests on both icf_model core
concerns, and on each of the constraint pool types, refile some large chunks of
icf_model_test.cc. Move the test helper setup methods into the new
icf_model_test_helpers, and move the constraint-focused tests into (some new,
some existing) *_constraints_pool_test files.

It should be reasonable to verify that aside from formalities, (like includes,
namespaces, etc.) only whole functions or whole test cases have moved. The test
case names haven't even changed, though that might get touched up in the
future.
